Move retryable lookup code into utils
Move the lookup code with retry logic into the utils code so that both the
integration and the End2End test can use them.

Add `lookup.go` with the code from the End2End package.

Add deprecation details to functions that have the same goal.

Update imports where possible and or required.
